[+] move function to function folder
