+++ includes/types/objects/Number/PrototypeProperties.hpp
@@ -1,5 +1,5 @@
-#ifndef JS_CMP_PROTOTYPEPROPERTIES_HPP
-#define JS_CMP_PROTOTYPEPROPERTIES_HPP
+#ifndef JS_CMP_PROTOTYPEPROPERTIESBOOLEAN_HPP
+#define JS_CMP_PROTOTYPEPROPERTIESBOOLEAN_HPP
 
 #include "internals/Object.hpp"
 
@@ -11,4 +11,4 @@ namespace JS::NUM {
 }
 
 
-#endif //JS_CMP_PROTOTYPEPROPERTIES_HPP
+#endif //JS_CMP_PROTOTYPEPROPERTIESBOOLEAN_HPP
