+++ lib/types/accessor.cpp
@@ -2,7 +2,7 @@
 #include "utils/Convert.hpp"
 
 #include <types/JsAny.hpp>
-#include <types/objects/JsFunction.hpp>
+#include "types/objects/Function/JsFunction.hpp"
 
 JS::PropertyProxy JS::Any::operator[](const std::string& key) const {
     if (this->value.index() == OBJECT) {
