+++ lib/types/objects/Object/constructor.cpp
@@ -1,4 +1,4 @@
-#include "types/objects/JsFunction.hpp"
+#include "types/objects/Function/JsFunction.hpp"
 #include "types/objects/Object/JsObject.hpp"
 #include "types/objects/Object/PrototypeProperties.hpp"
 #include <utility>
