+++ lib/types/objects/Object/prototype.cpp
@@ -1,5 +1,5 @@
 #include "types/objects/Object/PrototypeMethods.hpp"
-#include "types/objects/JsFunction.hpp"
+#include "types/objects/Function/JsFunction.hpp"
 #include "utils/Compare.hpp"
 #include "utils/Convert.hpp"
 #include "utils/Is.hpp"
