+++ lib/types/objects/Object/static.cpp
@@ -1,6 +1,6 @@
 #include "internals/Object.hpp"
 #include "types/objects/Object/StaticMethods.hpp"
-#include "types/objects/JsFunction.hpp"
+#include "types/objects/Function/JsFunction.hpp"
 #include "utils/Compare.hpp"
 #include "utils/Convert.hpp"
 #include "utils/Is.hpp"
