+++ lib/types/objects/String/prototype.cpp
@@ -1,4 +1,4 @@
-#include "types/objects/JsFunction.hpp"
+#include "types/objects/Function/JsFunction.hpp"
 #include "types/objects/String/PrototypeMethods.hpp"
 
 namespace JS::STR {
