+++ lib/types/operators/Modulo.cpp
@@ -12,7 +12,7 @@ JS::Any JS::Any::operator%(const JS::Any& other) const {
                     case STRING:
                         return JS::Any(std::fmod(std::get<double>(this->value),
                                                  JS::CONVERT::ToNumber(std::get<Rope>(other.value))));
-                    case BOOL:
+                    case BOOLEAN:
                         return JS::Any(
                             std::fmod(std::get<double>(this->value), static_cast<double>(std::get<bool>(other.value))));
                     default:
@@ -26,13 +26,13 @@ JS::Any JS::Any::operator%(const JS::Any& other) const {
                     case STRING:
                         return JS::Any(std::fmod(JS::CONVERT::ToNumber(std::get<Rope>(this->value)),
                                                  JS::CONVERT::ToNumber(std::get<Rope>(other.value))));
-                    case BOOL:
+                    case BOOLEAN:
                         return JS::Any(std::fmod(JS::CONVERT::ToNumber(std::get<Rope>(this->value)),
                                                  static_cast<double>(std::get<bool>(other.value))));
                     default:
                         return JS::Any(std::numeric_limits<double>::quiet_NaN());
                 }
-            case BOOL:
+            case BOOLEAN:
                 switch (other.value.index()) {
                     case NUMBER:
                         return JS::Any(
@@ -40,7 +40,7 @@ JS::Any JS::Any::operator%(const JS::Any& other) const {
                     case STRING:
                         return JS::Any(std::fmod(static_cast<double>(std::get<bool>(this->value)),
                                                  JS::CONVERT::ToNumber(std::get<Rope>(other.value))));
-                    case BOOL:
+                    case BOOLEAN:
                         return JS::Any(std::fmod(static_cast<double>(std::get<bool>(this->value)),
                                                  static_cast<double>(std::get<bool>(other.value))));
                     default:
@@ -52,7 +52,7 @@ JS::Any JS::Any::operator%(const JS::Any& other) const {
                         return JS::Any(std::fmod(0, std::get<double>(other.value)));
                     case STRING:
                         return JS::Any(std::fmod(0, JS::CONVERT::ToNumber(std::get<Rope>(other.value))));
-                    case BOOL:
+                    case BOOLEAN:
                         return JS::Any(std::fmod(0, static_cast<double>(std::get<bool>(other.value))));
                     default:
                         return JS::Any(std::numeric_limits<double>::quiet_NaN());
@@ -71,7 +71,7 @@ JS::Any JS::Any::operator%(int value) const {
             case STRING:
                 return JS::Any(
                     std::fmod(JS::CONVERT::ToNumber(std::get<Rope>(this->value)), static_cast<double>(value)));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(static_cast<double>(std::get<bool>(this->value)), static_cast<double>(value)));
             case NULL_TYPE:
                 return JS::Any(std::fmod(0, static_cast<double>(value)));
@@ -88,7 +88,7 @@ JS::Any JS::Any::operator%(double value) const {
                 return JS::Any(std::fmod(std::get<double>(this->value), value));
             case STRING:
                 return JS::Any(std::fmod(JS::CONVERT::ToNumber(std::get<Rope>(this->value)), value));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(static_cast<double>(std::get<bool>(this->value)), value));
             case NULL_TYPE:
                 return JS::Any(std::fmod(0, value));
@@ -106,7 +106,7 @@ JS::Any JS::Any::operator%(const char* value) const {
             case STRING:
                 return JS::Any(
                     std::fmod(JS::CONVERT::ToNumber(std::get<Rope>(this->value)), JS::CONVERT::ToNumber(value)));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(
                     std::fmod(static_cast<double>(std::get<bool>(this->value)), JS::CONVERT::ToNumber(value)));
             case NULL_TYPE:
@@ -125,7 +125,7 @@ JS::Any JS::Any::operator%(bool value) const {
             case STRING:
                 return JS::Any(
                     std::fmod(JS::CONVERT::ToNumber(std::get<Rope>(this->value)), static_cast<double>(value)));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(static_cast<double>(std::get<bool>(this->value)), static_cast<double>(value)));
             case NULL_TYPE:
                 return JS::Any(std::fmod(0, static_cast<double>(value)));
@@ -142,7 +142,7 @@ JS::Any JS::Any::operator%(JS::Null) const {
                 return JS::Any(std::fmod(std::get<double>(this->value), 0));
             case STRING:
                 return JS::Any(std::fmod(JS::CONVERT::ToNumber(std::get<Rope>(this->value)), 0));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(static_cast<double>(std::get<bool>(this->value)), 0));
             default:
                 return JS::Any(std::numeric_limits<double>::quiet_NaN());
@@ -158,7 +158,7 @@ JS::Any JS::Any::operator%(JS::Undefined) const {
             case STRING:
                 return JS::Any(std::fmod(JS::CONVERT::ToNumber(std::get<Rope>(this->value)),
                                          std::numeric_limits<double>::quiet_NaN()));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(static_cast<double>(std::get<bool>(this->value)),
                                          std::numeric_limits<double>::quiet_NaN()));
             default:
@@ -175,7 +175,7 @@ Any operator%(int value, JS::Any const& any) {
                 return JS::Any(std::fmod(value, std::get<double>(any.value)));
             case STRING:
                 return JS::Any(std::fmod(static_cast<double>(value), JS::CONVERT::ToNumber(std::get<Rope>(any.value))));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(static_cast<double>(value), static_cast<double>(std::get<bool>(any.value))));
             default:
                 return JS::Any(std::numeric_limits<double>::quiet_NaN());
@@ -190,7 +190,7 @@ Any operator%(double value, JS::Any const& any) {
                 return JS::Any(std::fmod(value, std::get<double>(any.value)));
             case STRING:
                 return JS::Any(std::fmod(value, JS::CONVERT::ToNumber(std::get<Rope>(any.value))));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(value, static_cast<double>(std::get<bool>(any.value))));
             default:
                 return JS::Any(std::numeric_limits<double>::quiet_NaN());
@@ -206,7 +206,7 @@ Any operator%(const char* value, JS::Any const& any) {
             case STRING:
                 return JS::Any(
                     std::fmod(JS::CONVERT::ToNumber(value), JS::CONVERT::ToNumber(std::get<Rope>(any.value))));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(JS::CONVERT::ToNumber(value), static_cast<double>(std::get<bool>(any.value))));
             default:
                 return JS::Any(std::numeric_limits<double>::quiet_NaN());
@@ -221,7 +221,7 @@ Any operator%(bool value, JS::Any const& any) {
                 return JS::Any(std::fmod(static_cast<double>(value), std::get<double>(any.value)));
             case STRING:
                 return JS::Any(std::fmod(static_cast<double>(value), JS::CONVERT::ToNumber(std::get<Rope>(any.value))));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(static_cast<double>(value), static_cast<double>(std::get<bool>(any.value))));
             default:
                 return JS::Any(std::numeric_limits<double>::quiet_NaN());
@@ -236,7 +236,7 @@ Any operator%(JS::Null value, JS::Any const& any) {
                 return JS::Any(std::fmod(0, std::get<double>(any.value)));
             case STRING:
                 return JS::Any(std::fmod(0, JS::CONVERT::ToNumber(std::get<Rope>(any.value))));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(0, static_cast<double>(std::get<bool>(any.value))));
             default:
                 return JS::Any(std::numeric_limits<double>::quiet_NaN());
@@ -252,7 +252,7 @@ Any operator%(JS::Undefined value, JS::Any const& any) {
             case STRING:
                 return JS::Any(std::fmod(std::numeric_limits<double>::quiet_NaN(),
                                          JS::CONVERT::ToNumber(std::get<Rope>(any.value))));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::fmod(std::numeric_limits<double>::quiet_NaN(),
                                          static_cast<double>(std::get<bool>(any.value))));
             default:
