+++ lib/types/operators/Mult.cpp
@@ -13,7 +13,7 @@ JS::Any JS::Any::operator*(const JS::Any& other) const {
                     case STRING:
                         return JS::Any(std::get<double>(this->value) *
                                        JS::CONVERT::ToNumber(std::get<Rope>(other.value)));
-                    case BOOL:
+                    case BOOLEAN:
                         return JS::Any(std::get<double>(this->value) * std::get<bool>(other.value));
                     case NULL_TYPE:
                         return JS::Any(0);
@@ -29,7 +29,7 @@ JS::Any JS::Any::operator*(const JS::Any& other) const {
                     case STRING:
                         return JS::Any(JS::CONVERT::ToNumber(std::get<Rope>(this->value)) *
                                        JS::CONVERT::ToNumber(std::get<Rope>(other.value)));
-                    case BOOL:
+                    case BOOLEAN:
                         return JS::Any(JS::CONVERT::ToNumber(std::get<Rope>(this->value)) *
                                        std::get<bool>(other.value));
                     case NULL_TYPE:
@@ -38,7 +38,7 @@ JS::Any JS::Any::operator*(const JS::Any& other) const {
                     default:
                         return JS::Any(std::numeric_limits<double>::quiet_NaN());
                 }
-            case BOOL:
+            case BOOLEAN:
 
                 switch (other.value.index()) {
                     case NUMBER:
@@ -46,7 +46,7 @@ JS::Any JS::Any::operator*(const JS::Any& other) const {
                     case STRING:
                         return JS::Any(std::get<bool>(this->value) *
                                        JS::CONVERT::ToNumber(std::get<Rope>(other.value)));
-                    case BOOL:
+                    case BOOLEAN:
                         return JS::Any(std::get<bool>(this->value) * std::get<bool>(other.value));
                     case NULL_TYPE:
                         return JS::Any(0);
@@ -60,7 +60,7 @@ JS::Any JS::Any::operator*(const JS::Any& other) const {
                     case STRING:
                         JS::CONVERT::ToNumber(std::get<Rope>(other.value));
                         return JS::Any(0);
-                    case BOOL:
+                    case BOOLEAN:
                         return JS::Any(0);
                     case NULL_TYPE:
                         return JS::Any(0);
@@ -80,7 +80,7 @@ JS::Any JS::Any::operator*(int value) const {
                 return JS::Any(std::get<double>(this->value) * value);
             case STRING:
                 return JS::Any(JS::CONVERT::ToNumber(std::get<Rope>(this->value)) * value);
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::get<bool>(this->value) * value);
             case NULL_TYPE:
                 return JS::Any(0.0);
@@ -97,7 +97,7 @@ JS::Any JS::Any::operator*(double value) const {
                 return JS::Any(std::get<double>(this->value) * value);
             case STRING:
                 return JS::Any(JS::CONVERT::ToNumber(std::get<Rope>(this->value)) * value);
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::get<bool>(this->value) * value);
             case NULL_TYPE:
                 return JS::Any(0);
@@ -114,7 +114,7 @@ JS::Any JS::Any::operator*(const char* value) const {
                 return JS::Any(std::get<double>(this->value) * JS::CONVERT::ToNumber(value));
             case STRING:
                 return JS::Any(JS::CONVERT::ToNumber(std::get<Rope>(this->value)) * JS::CONVERT::ToNumber(value));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::get<bool>(this->value) * JS::CONVERT::ToNumber(value));
             case NULL_TYPE: {
                 JS::CONVERT::ToNumber(value);
@@ -133,7 +133,7 @@ JS::Any JS::Any::operator*(bool value) const {
                 return JS::Any(std::get<double>(this->value) * static_cast<double>(value));
             case STRING:
                 return JS::Any(JS::CONVERT::ToNumber(std::get<Rope>(this->value)) * static_cast<double>(value));
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::get<bool>(this->value) * value);
             case NULL_TYPE:
                 return JS::Any(0);
@@ -150,7 +150,7 @@ JS::Any JS::Any::operator*(JS::Null) const {
                 return JS::Any(std::get<double>(this->value) * 0.0);
             case STRING:
                 return JS::Any(JS::CONVERT::ToNumber(std::get<Rope>(this->value)) * 0.0);
-            case BOOL:
+            case BOOLEAN:
                 return JS::Any(std::get<bool>(this->value) * 0.0);
             case NULL_TYPE:
                 return JS::Any(0);
@@ -170,7 +170,7 @@ Any operator*(int value, const JS::Any& any) {
                 return JS::Any(value * std::get<double>(any.getValue()));
             case JS::STRING:
                 return JS::Any(value * JS::CONVERT::ToNumber(std::get<Rope>(any.getValue())));
-            case JS::BOOL:
+            case JS::BOOLEAN:
                 return JS::Any(value * std::get<bool>(any.getValue()));
             case JS::NULL_TYPE:
                 return JS::Any(0);
@@ -187,7 +187,7 @@ Any operator*(double value, const JS::Any& any) {
                 return JS::Any(value * std::get<double>(any.getValue()));
             case JS::STRING:
                 return JS::Any(value * JS::CONVERT::ToNumber(std::get<Rope>(any.getValue())));
-            case JS::BOOL:
+            case JS::BOOLEAN:
                 return JS::Any(value * std::get<bool>(any.getValue()));
             case JS::NULL_TYPE:
                 return JS::Any(0);
@@ -204,7 +204,7 @@ Any operator*(const char* value, const JS::Any& any) {
                 return JS::Any(JS::CONVERT::ToNumber(value) * std::get<double>(any.getValue()));
             case JS::STRING:
                 return JS::Any(JS::CONVERT::ToNumber(value) * JS::CONVERT::ToNumber(std::get<Rope>(any.getValue())));
-            case JS::BOOL:
+            case JS::BOOLEAN:
                 return JS::Any(JS::CONVERT::ToNumber(value) * std::get<bool>(any.getValue()));
             case JS::NULL_TYPE:
                 JS::CONVERT::ToNumber(value);
@@ -222,7 +222,7 @@ Any operator*(bool value, const JS::Any& any) {
                 return JS::Any(static_cast<double>(value) * std::get<double>(any.getValue()));
             case JS::STRING:
                 return JS::Any(static_cast<double>(value) * JS::CONVERT::ToNumber(std::get<Rope>(any.getValue())));
-            case JS::BOOL:
+            case JS::BOOLEAN:
                 return JS::Any(value * std::get<bool>(any.getValue()));
             case JS::NULL_TYPE:
                 return JS::Any(0);
@@ -239,7 +239,7 @@ Any operator*(JS::Null, const JS::Any& any) {
                 return JS::Any(0.0 * std::get<double>(any.getValue()));
             case JS::STRING:
                 return JS::Any(0.0 * JS::CONVERT::ToNumber(std::get<Rope>(any.getValue())));
-            case JS::BOOL:
+            case JS::BOOLEAN:
                 return JS::Any(0.0 * std::get<bool>(any.getValue()));
             case JS::NULL_TYPE:
                 return JS::Any(0);
